requirements.txt
fastapi
uvicorn
sqlalchemy[psycopg2]
pydantic[email]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]